+++ core/shared/src/main/scala/zio/logging/ConfigurableLogger.scala
@@ -1,5 +1,21 @@
+/*
+ * Copyright 2019-2023 John A. De Goes and the ZIO Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package zio.logging
 
+import zio.logging.internal.ReconfigurableLogger
 import zio.{ Cause, FiberId, FiberRef, FiberRefs, LogLevel, LogSpan, Trace, ZIO, ZLayer, ZLogger }
 
 trait LoggerConfigurer {
@@ -49,13 +65,43 @@ trait ConfigurableLogger[-Message, +Output] extends ZLogger[Message, Output] {
 
 object ConfigurableLogger {
 
-  def apply[Message, Output](
-    logger: ZLogger[Message, Output],
-    loggerConfigurer: LoggerConfigurer
-  ): ConfigurableLogger[Message, Output] =
-    new ConfigurableLogger[Message, Output] {
+//  def apply[Message, Output](
+//    logger: ZLogger[Message, Output],
+//    loggerConfigurer: LoggerConfigurer
+//  ): ConfigurableLogger[Message, Output] =
+//    new ConfigurableLogger[Message, Output] {
+//
+//      override val configurer: LoggerConfigurer = loggerConfigurer
+//
+//      override def apply(
+//        trace: Trace,
+//        fiberId: FiberId,
+//        logLevel: LogLevel,
+//        message: () => Message,
+//        cause: Cause[Any],
+//        context: FiberRefs,
+//        spans: List[LogSpan],
+//        annotations: Map[String, String]
+//      ): Output = logger.apply(trace, fiberId, logLevel, message, cause, context, spans, annotations)
+//    }
 
-      override val configurer: LoggerConfigurer = loggerConfigurer
+  def make[Message, Output](
+    logger: ZLogger[Message, Output],
+    filterConfig: LogFilter.LogLevelByNameConfig
+  ): ConfigurableLogger[Message, Option[Output]] = {
+
+    val reconfigurableLogger = ReconfigurableLogger[Message, Option[Output], LogFilter.LogLevelByNameConfig](
+      filterConfig,
+      config => {
+        val filter = LogFilter.logLevelByName(config)
+
+        filter.filter(logger)
+      }
+    )
+
+    new ConfigurableLogger[Message, Option[Output]] {
+
+      override val configurer: LoggerConfigurer = Configurer(reconfigurableLogger)
 
       override def apply(
         trace: Trace,
@@ -66,6 +112,52 @@ object ConfigurableLogger {
         context: FiberRefs,
         spans: List[LogSpan],
         annotations: Map[String, String]
-      ): Output = logger.apply(trace, fiberId, logLevel, message, cause, context, spans, annotations)
+      ): Option[Output] =
+        reconfigurableLogger.apply(trace, fiberId, logLevel, message, cause, context, spans, annotations)
     }
+  }
+
+  private case class Configurer(logger: ReconfigurableLogger[_, _, LogFilter.LogLevelByNameConfig])
+      extends LoggerConfigurer {
+
+    private val rootName = "root"
+
+    override def getLoggerConfigs(): ZIO[Any, Throwable, List[LoggerConfigurer.LoggerConfig]] =
+      ZIO.attempt {
+        val currentConfig = logger.config
+
+        LoggerConfigurer.LoggerConfig(rootName, currentConfig.rootLevel) :: currentConfig.mappings
+          .map(LoggerConfigurer.LoggerConfig.tupled)
+          .toList
+      }
+
+    override def getLoggerConfig(name: String): ZIO[Any, Throwable, Option[LoggerConfigurer.LoggerConfig]] =
+      ZIO.attempt {
+        val currentConfig = logger.config
+
+        if (name == rootName) {
+          Some(LoggerConfigurer.LoggerConfig(rootName, currentConfig.rootLevel))
+        } else {
+          currentConfig.mappings.collectFirst {
+            case (n, l) if n == name => LoggerConfigurer.LoggerConfig(n, l)
+          }
+        }
+      }
+
+    override def setLoggerConfig(name: String, logLevel: LogLevel): ZIO[Any, Throwable, LoggerConfigurer.LoggerConfig] =
+      ZIO.attempt {
+        val currentConfig = logger.config
+
+        val newConfig = if (name == rootName) {
+          currentConfig.withRootLevel(logLevel)
+        } else {
+          currentConfig.withMapping(name, logLevel)
+        }
+
+        logger.reconfigureIfChanged(newConfig)
+
+        LoggerConfigurer.LoggerConfig(name, logLevel)
+      }
+  }
+
 }
